1. DAL completed 2. Added scrollers for long descriptions
