+++ WebProject/Services/CommentService.cs
@@ -1,5 +1,4 @@
-﻿
-using Microsoft.AspNetCore.SignalR;
+﻿using Microsoft.AspNetCore.SignalR;
 using WebProject.Hubs;
 
 namespace WebProject.Services
