+++ WebProject/Views/Account/LoginForm.cshtml
@@ -1,4 +1,4 @@
-﻿@model WebProject.Models.UserModel
+﻿@model ZooLib.Models.UserModel
 
 @{
     ViewBag.Title = "Login";
