+++ WebProject/Views/Administrator/AddAnimalForm.cshtml
@@ -1,4 +1,4 @@
-﻿@model WebProject.Models.AnimalViewModel
+﻿@model ZooLib.Models.AnimalViewModel
 
 @{
     ViewBag.Title = "AddAnimalForm";
