+++ WebProject/Views/Administrator/DeleteAnimalForm.cshtml
@@ -1,4 +1,4 @@
-﻿@model WebProject.Models.Animal;
+﻿@model ZooLib.Models.Animal;
 
 @{
     ViewBag.Title = "DeleteAnimal";
