+++ WebProject/Views/Administrator/EditAnimalForm.cshtml
@@ -1,4 +1,4 @@
-﻿@model WebProject.Models.AnimalModelEditor
+﻿@model ZooLib.Models.AnimalModelEditor
 
 @{
     ViewBag.Title = "EditAnimalForm";
