+++ WebProject/Views/Catalog/Catalog.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<WebProject.Models.Animal>
+﻿@model IEnumerable<ZooLib.Models.Animal>
 
 @{
     ViewBag.Title = "Catalog";
