+++ WebProject/Views/Home/Index.cshtml
@@ -1,4 +1,4 @@
-﻿@model IEnumerable<WebProject.Models.Animal>
+﻿@model IEnumerable<ZooLib.Models.Animal>
 
 @{
     ViewBag.Title = "Home";
@@ -67,7 +67,7 @@ else
             <div class="col-md-9">
                 <div class="card-body">
                     <h5 class="card-title">@a.Name</h5>
-                    <p class="card-text">@a.Description</p>
+                    <p class="card-text scroll-container">@a.Description</p>
                     <p class="card-text"><small class="text-body-secondary">Number of comments: @a.Comments!.Count</small></p>
                 </div>
             </div>
