+++ WebProject/WebProject.csproj
@@ -15,4 +15,8 @@
     <PackageReference Include="Serilog.AspNetCore" Version="8.0.2" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\ZooLib\ZooLib.csproj" />
+  </ItemGroup>
+
 </Project>
