+++ ZooLib/Data/SeedData.cs
@@ -1,6 +1,6 @@
 ﻿using Microsoft.AspNetCore.Identity;
 
-namespace WebProject.Data
+namespace ZooLib.Data
 {
     public class SeedData
     {
